requirements.txt
matplotlib==3.1.2
mediapipe==0.8.9
numpy==1.17.4
opencv_python==4.5.4.60
phys==0.1.0
PyQt5==5.15.6
roboticstoolbox_python==0.11.0
rosgraph==1.15.14
rospy==1.15.14
sensor_msgs==1.13.1
spatialgeometry==0.2.0
spatialmath_python==0.11
swift==2.29.1
swift_sim==0.10.0
websockets==10.0
